+++ src/main/java/ukma/groupproject/shop/service/impl/SupplierServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import ukma.groupproject.shop.dao.OrderDao;
 import ukma.groupproject.shop.dao.SupplierDao;
 import ukma.groupproject.shop.model.Item;
 import ukma.groupproject.shop.model.Supplier;
@@ -21,16 +22,14 @@ public class SupplierServiceImpl implements SupplierService {
     private SupplierDao supplierDao;
 
     @Autowired
+    private OrderDao orderDao;
+
+    @Autowired
     private ItemService itemService;
 
     @Override
     public Supplier get(Long id) {
         return supplierDao.get(id);
-    }
-
-    @Override
-    public Supplier getByName(String name) {
-        return supplierDao.getByName(name);
     }
 
     @Override
@@ -51,14 +50,26 @@ public class SupplierServiceImpl implements SupplierService {
 
     @Override
 	public void persist(Supplier s) {
-        if(getByName(s.getName()) != null) {
+        if(supplierDao.getByName(s.getName()) != null) {
             throw new ShopBusinessException("Supplier with specified name already exists");
         }
 		supplierDao.persist(s);
 	}
 
-	@Override
+    @Override
+    public void update(Supplier s) {
+        Supplier byName = supplierDao.getByName(s.getName());
+        if(byName != null && !byName.equals(s)) {
+            throw new ShopBusinessException("Supplier with specified name already exists");
+        }
+        supplierDao.update(s);
+    }
+
+    @Override
 	public void delete(Supplier s) {
+        if(orderDao.existOrdersFor(s)) {
+            throw new ShopBusinessException("Supplier already has active orders");
+        }
         supplierDao.delete(s);
 	}
 }
